Extract checker to own file
