+++ src/main/java/net/sf/jabref/logic/integrity/IntegrityCheck.java
@@ -340,51 +340,6 @@ public class IntegrityCheck {
         }
     }
 
-    private static class EditionChecker implements Checker {
-
-        private static final Predicate<String> FIRST_LETTER_CAPITALIZED = Pattern.compile("^[A-Z]").asPredicate();
-        private static final Predicate<String> ONLY_NUMERALS_OR_LITERALS = Pattern.compile("^([0-9]+|[^0-9]+)$")
-                .asPredicate();
-
-        private final BibDatabaseContext bibDatabaseContextEdition;
-
-        public EditionChecker(BibDatabaseContext bibDatabaseContext) {
-            this.bibDatabaseContextEdition = Objects.requireNonNull(bibDatabaseContext);
-        }
-
-        /**
-         * Checks, if the first letter is capitalized (BibTeX mode)
-         * Checks, if field contains only an integer (BibLaTeX mode)
-         * Official bibtex spec:
-         * The edition of a book-for example, “Second”.
-         * This should be an ordinal, and should have the first letter capitalized.
-         * BibLaTex:
-         * The edition of a printed publication. This must be an integer, not an ordinal.
-         * It is also possible to give the edition as a literal string, for example "Third, revised and expanded edition".
-         */
-        @Override
-        public List<IntegrityMessage> check(BibEntry entry) {
-            Optional<String> value = entry.getField(FieldName.EDITION);
-            if (!value.isPresent()) {
-                return Collections.emptyList();
-            }
-
-            //BibLaTeX
-            if (!ONLY_NUMERALS_OR_LITERALS.test(value.get().trim()) && (bibDatabaseContextEdition.isBiblatexMode())) {
-                return Collections.singletonList(new IntegrityMessage(
-                        Localization.lang("should contain an integer or a literal"), entry, FieldName.EDITION));
-            }
-
-            //BibTeX
-            if (!FIRST_LETTER_CAPITALIZED.test(value.get().trim()) && (!bibDatabaseContextEdition.isBiblatexMode())) {
-                return Collections.singletonList(new IntegrityMessage(
-                        Localization.lang("should have the first letter capitalized"), entry, FieldName.EDITION));
-            }
-
-            return Collections.emptyList();
-        }
-    }
-
     /**
      * From BibTex manual:
      * One or more page numbers or range of numbers, such as 42--111 or 7,41,73--97 or 43+
